+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* On first visit, DIM will prompt you to select a platform instead of automatically selecting the most recently played one. Also, DIM will no longer fall back to your D1 account when Bungie.net is down.
 * Invalid search queries are now detected more reliably and DIM will not show search results if the query is invalid.
 * Loadout Optimizer will now remember stat priorities and enabled stats per Guardian class.